requirements.txt
torch
faiss-cpu
sentence-transformers
transformers
bitsandbytes
accelerate
pypdf
python-docx
